chore(api): extract gcs_client.py from uploads.py (#195)
Moves all GCS plumbing (get_gcp_access_token, upload_to_gcs,
_generate_signed_url_v4, _sign_string_with_iam, _fetch_metadata,
_encode_rfc3986, _encode_path, GCS_API_BASE) into a dedicated
src/gcs_client.py module. uploads.py re-exports them; multipart_uploads.py
imports from gcs_client directly, removing the legacy_uploads coupling.
Test patch paths updated accordingly.
